Refactor : Service instead using DAO directly Closes #55
