+++ src/main/java/app/rl/blog/service/PostService.java
@@ -1,12 +1,15 @@
-package app.rl.blog.dao;
+package app.rl.blog.service;
 
 import java.util.List;
+
+import org.springframework.stereotype.Service;
 
 import app.rl.blog.entity.Post;
 import jakarta.validation.Valid;
 
-public interface PostDAO {
-
+@Service
+public interface PostService {
+    
     public Post savePost(@Valid Post post);
 
     public List<Post> fetchPostList();
